Add extensible key map to configuration
