+++ src/Termonad/Keys.hs
@@ -3,23 +3,11 @@ module Termonad.Keys where
 
 import Termonad.Prelude
 
-import Control.Lens (imap)
-import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import Data.Set (Set)
 import qualified Data.Set as Set
 import GI.Gdk
   ( EventKey
-  , pattern KEY_0
-  , pattern KEY_1
-  , pattern KEY_2
-  , pattern KEY_3
-  , pattern KEY_4
-  , pattern KEY_5
-  , pattern KEY_6
-  , pattern KEY_7
-  , pattern KEY_8
-  , pattern KEY_9
   , ModifierType(..)
   , getEventKeyHardwareKeycode
   , getEventKeyIsModifier
@@ -30,9 +18,14 @@ import GI.Gdk
   , getEventKeyType
   )
 
-import Termonad.Term (altNumSwitchTerm)
-import Termonad.Types (TMState, TMWindowId)
-
+import Termonad.Types
+  ( Key(..)
+  , TMState
+  , TMState'(tmStateConfig)
+  , TMWindowId
+  , TMConfig(keys)
+  , toKey
+  )
 
 showKeys :: EventKey -> IO Bool
 showKeys eventKey = do
@@ -55,42 +48,6 @@ showKeys eventKey = do
   putStrLn ""
 
   pure True
-
-data Key = Key
-  { keyVal :: !Word32
-  , keyMods :: !(Set ModifierType)
-  } deriving (Eq, Ord, Show)
-
-toKey :: Word32 -> Set ModifierType -> Key
-toKey = Key
-
-keyMap :: Map Key (TMState -> TMWindowId -> IO Bool)
-keyMap =
-  let numKeys :: [Word32]
-      numKeys =
-        [ KEY_1
-        , KEY_2
-        , KEY_3
-        , KEY_4
-        , KEY_5
-        , KEY_6
-        , KEY_7
-        , KEY_8
-        , KEY_9
-        , KEY_0
-        ]
-      altNumKeys :: [(Key, TMState -> TMWindowId -> IO Bool)]
-      altNumKeys =
-        imap
-          (\i k ->
-             (toKey k [ModifierTypeMod1Mask], stopProp (altNumSwitchTerm i))
-          )
-          numKeys
-  in
-  Map.fromList altNumKeys
-
-stopProp :: (TMState -> TMWindowId -> IO a) -> TMState -> TMWindowId -> IO Bool
-stopProp callback terState tmWinId = callback terState tmWinId $> True
 
 removeStrangeModifiers :: Key -> Key
 removeStrangeModifiers Key{keyVal, keyMods} =
@@ -121,7 +78,7 @@ handleKeyPress terState tmWindowId eventKey = do
   modifiers <- getEventKeyState eventKey
   let oldKey = toKey keyval (Set.fromList modifiers)
       newKey = removeStrangeModifiers oldKey
-      maybeAction = Map.lookup newKey keyMap
+  maybeAction <- Map.lookup newKey . keys . tmStateConfig <$> readMVar terState
   case maybeAction of
     Just action -> action terState tmWindowId
     Nothing -> pure False
